backend/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
Authlib==1.3.2
certifi==2024.8.30
cffi==1.17.1
click==8.1.7
cryptography==43.0.1
dnspython==2.6.1
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.114.0
fastapi-cli==0.0.5
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.2
idna==3.8
itsdangerous==2.2.0
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
orjson==3.10.7
pycparser==2.22
pydantic==2.9.0
pydantic-extra-types==2.9.0
pydantic-settings==2.4.0
pydantic_core==2.23.2
Pygments==2.18.0
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.2
rich==13.8.0
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==2.0.34
starlette==0.38.5
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
uvicorn==0.30.6
uvloop==0.20.0
watchfiles==0.24.0
websockets==13.0.1
PyJWT
passlib
bcrypt
psycopg2-binary
python-jose
